+++ lib/plivo/base/resource_interface.rb
@@ -59,6 +59,13 @@ module Plivo
         )
       end
 
+      def perform_create_with_id(id, params, use_multipart_conn=false)
+          Response.new(
+            @_client.send_request(@_resource_uri + id + '/', 'POST', params, 10, use_multipart_conn, is_voice_request: @_is_voice_request),
+            @_identifier_string
+          )
+      end
+
       def perform_delete(identifier, params = nil)
         valid_param?(:identifier, identifier, [String, Symbol], true)
         Response.new(
@@ -97,7 +104,7 @@ module Plivo
           objects: @_resource_list
         }
       end
-      
+
       def perform_action(action = nil, method = 'GET', params = nil, parse = false)
         resource_path = action ? @_resource_uri + action + '/' : @_resource_uri
         response = @_client.send_request(resource_path, method, params, nil, false, is_voice_request: @_is_voice_request)